+++ LittleWarGame/Rescue.cs
@@ -22,7 +22,7 @@ namespace LittleWarGame
 
             CDTime.setCoolDownTime(15);
 
-            img.Image = myRealStatus[Const.Status.move];
+            img.Image = myRealStatus[(int)Const.Status.move];
             img.Top = Const.mainLineHeight - Const.warriorHeight;
         }
         public override void helpTo(Warriors we)
